Explanation to user
On the drop-down list, lets add an option called "Plot your own data". Then in the white box, give instructions (with the # symbol so they are commented out). eg. tell them they can paste in from a spreadsheet. Tell them the top row is expected to be a header line. Tell them the data can be tab-separated or comma-separated. And anything else explaining how to use the tool.